Feature: Update organization endpoint

+++ ClockiGo.Application/CQRS/Commands/Organization/DeleteOrganizationCommand/DeleteOrganizationCommandHandler.cs
@@ -9,6 +9,7 @@ namespace ClockiGo.Application.CQRS.Commands.Organization.DeleteOrganizationComm
     public class DeleteOrganizationCommandHandler : IRequestHandler<DeleteOrganizationCommand, ErrorOr<DeleteOrganizationResult>>
     {
         private readonly IOrganizationRepository _organizationRepository;
+        private readonly IUserRepository _userRepository;
 
         public DeleteOrganizationCommandHandler(IOrganizationRepository organizationRepository)
         {
@@ -17,6 +18,11 @@ namespace ClockiGo.Application.CQRS.Commands.Organization.DeleteOrganizationComm
 
         public async Task<ErrorOr<DeleteOrganizationResult>> Handle(DeleteOrganizationCommand request, CancellationToken cancellationToken)
         {
+            var user = await _userRepository.GetUserByIdAsync(request.UserId);
+            if (user is null) return Errors.User.UserNotFound;
+            if (user.Role != Domain.Enums.Role.Admin) return Errors.User.AccessDenied;
+
+
             var isSuccess = await _organizationRepository.DeleteAsync(request.OrganizationId);
 
             if (!isSuccess)
